devops-pipelines/load-testing/latency/requirements.txt
azure-eventhub==5.7.0
pytest
pytest-asyncio